requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
mysql_connector_repackaged==0.3.1
